+++ src/slog/LogSink.hpp
@@ -39,12 +39,12 @@ class NullSink : public LogSink
  * This should write rec to sink and return the number of bytes
  * written.
  */
-using Formatter = std::function<uint32_t(FILE* sink, LogRecord const& node)>;
+using Formatter = std::function<long(FILE* sink, LogRecord const& node)>;
 
 /**
  * @brief Write data to a log file before/after records are added
  */
-using LogFileFurniture = std::function<int(FILE* sink, int sequence, uint64_t time)>;
+using LogFileFurniture = std::function<long(FILE* sink, int sequence, uint64_t time)>;
 
 /**
  * @brief Writes the severity level as a four character string to severity_str
@@ -99,7 +99,7 @@ uint32_t total_record_size(LogRecord const& node);
  * Using the byte count, write rec to the file. Follow all of the more()
  * pointers as required.
  */
-uint32_t write_message_to_file(FILE* sink, LogRecord const& rec);
+long write_message_to_file(FILE* sink, LogRecord const& rec);
 
 /**
  * @brief The default record format: "[SEVR TAG YYYY-MM-DD hh:mm:ss.sssZ]"
@@ -107,12 +107,12 @@ uint32_t write_message_to_file(FILE* sink, LogRecord const& rec);
  * SEVR is the four character severity code from format_severity(), TAG is the
  * optional tag (if no tag, then nothing is printed).
  */
-uint32_t default_format(FILE* sink, LogRecord const& node);
+long default_format(FILE* sink, LogRecord const& node);
 
 /**
  * A Formatter that just prints the records
  */
-uint32_t no_meta_format(FILE* sink, LogRecord const& node);
+long no_meta_format(FILE* sink, LogRecord const& node);
 
 /**
  * @brief This writes data in a byte delimited format with an 8 byte record header
@@ -124,7 +124,7 @@ uint32_t no_meta_format(FILE* sink, LogRecord const& node);
  *    <tag> is the first four bytes of the tag (short tags are padded out with nulls)
  *    <record> is the logged data
  */
-uint32_t default_binary_format(FILE* sink, LogRecord const& node);
+long default_binary_format(FILE* sink, LogRecord const& node);
 
 /**
  * @brief This writes data in a byte delimited format with a 32 byte record header.
@@ -139,18 +139,18 @@ uint32_t default_binary_format(FILE* sink, LogRecord const& node);
  *    <tag> is 16 one-byte characters. The final character is always '\0'
  *    <record> is the logged data
  */
-uint32_t long_binary_format(FILE* sink, LogRecord const& node);
+long long_binary_format(FILE* sink, LogRecord const& node);
 
 /**
  * @brief Simple 8 byte file header of the form "SLOG<BOM><SS>" where the literal
  * ascii characters "SLOG" start the file, followed by the two byte byte-order-mark
  * 0xFEFF, followed by the two byte sequence number. The time is not included.
  */
-uint32_t default_binary_header_furniture(FILE* sink, int sequence, uint64_t time);
+long default_binary_header_furniture(FILE* sink, int sequence, uint64_t time);
 
 /**
  * @brief A furniture function that writes nothing
  */
-inline uint32_t no_op_furniture(FILE*, int, uint64_t) { return 0; }
+inline long no_op_furniture(FILE*, int, uint64_t) { return 0; }
 
 } // namespace slog
